Add welcome + refactor
