+++ src/Arcus.EventGrid.Sidecar.Api/Extensions/IApplicationBuilderExtensions.cs
@@ -1,0 +1,26 @@
+﻿using Swashbuckle.AspNetCore.SwaggerUI;
+
+namespace Microsoft.AspNetCore.Builder
+{
+    public static class IApplicationBuilderExtensions
+    {
+        /// <summary>
+        /// Add support for Open API documentation with API explorer
+        /// </summary>
+        /// <param name="applicationBuilder">Application Builder</param>
+        public static void UseOpenApiDocsWithExplorer(this IApplicationBuilder applicationBuilder)
+        {
+            applicationBuilder.UseSwagger();
+            applicationBuilder.UseSwaggerUI(swaggerUiOptions =>
+            {
+                swaggerUiOptions.SwaggerEndpoint("/swagger/v1/swagger.json", "Arcus Event Grid v1 API");
+                swaggerUiOptions.DisplayOperationId();
+                swaggerUiOptions.EnableDeepLinking();
+                swaggerUiOptions.DocumentTitle = "Arcus Event Grid API";
+                swaggerUiOptions.DocExpansion(DocExpansion.List);
+                swaggerUiOptions.DisplayRequestDuration();
+                swaggerUiOptions.EnableFilter();
+            });
+        }
+    }
+}
