+++ src/Arcus.EventGrid.Sidecar.Api/Program.cs
@@ -1,13 +1,31 @@
-﻿using Microsoft.AspNetCore;
+﻿using System;
+using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 
 namespace Arcus.EventGrid.Sidecar.Api
 {
     public class Program
     {
+        private const string WelcomeText = @"
+ █████╗ ██████╗  ██████╗██╗   ██╗███████╗
+██╔══██╗██╔══██╗██╔════╝██║   ██║██╔════╝
+███████║██████╔╝██║     ██║   ██║███████╗
+██╔══██║██╔══██╗██║     ██║   ██║╚════██║
+██║  ██║██║  ██║╚██████╗╚██████╔╝███████║
+╚═╝  ╚═╝╚═╝  ╚═╝ ╚═════╝ ╚═════╝ ╚══════╝";
+
         public static void Main(string[] args)
         {
-            CreateWebHostBuilder(args).Build().Run();
+            Welcome();
+
+            CreateWebHostBuilder(args)
+                .Build()
+                .Run();
+        }
+
+        private static void Welcome()
+        {
+            Console.WriteLine(WelcomeText);
         }
 
         public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
