+++ src/Arcus.EventGrid.Sidecar.Api/Startup.cs
@@ -1,12 +1,8 @@
-﻿using System.IO;
-using System.Linq;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Swashbuckle.AspNetCore.Swagger;
-using Swashbuckle.AspNetCore.SwaggerUI;
 
 namespace Arcus.EventGrid.Sidecar.Api
 {
@@ -30,35 +26,7 @@ namespace Arcus.EventGrid.Sidecar.Api
                     })
                     .SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
 
-            var openApiInformation = new Info
-            {
-                Contact = new Contact
-                {
-                    Name = "Arcus",
-                    Url = "https://github.com/arcus-azure/arcus.eventgrid.sidecar"
-                },
-                Title = "Arcus",
-                Description = "Push events to Azure Event Grid in a breeze - API allows you to send events over REST without having to worry about how Azure Event Grid works.",
-                License = new License
-                {
-                    Name = "MIT",
-                    Url = "https://github.com/arcus-azure/arcus.eventgrid.sidecar/blob/master/LICENSE"
-                }
-            };
-
-            var xmlDocumentationPath = GetXmlDocumentationPath(services);
-
-            services.AddSwaggerGen(swaggerGenerationOptions =>
-            {
-                swaggerGenerationOptions.EnableAnnotations();
-                swaggerGenerationOptions.SwaggerDoc("v1", openApiInformation);
-                swaggerGenerationOptions.DescribeAllEnumsAsStrings();
-
-                if (string.IsNullOrEmpty(xmlDocumentationPath) == false)
-                {
-                    swaggerGenerationOptions.IncludeXmlComments(xmlDocumentationPath);
-                }
-            });
+            services.UseOpenApiSpecifications();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
@@ -76,31 +44,7 @@ namespace Arcus.EventGrid.Sidecar.Api
 
             app.UseHttpsRedirection();
             app.UseMvc();
-            app.UseSwagger();
-            app.UseSwaggerUI(swaggerUiOptions =>
-            {
-                swaggerUiOptions.SwaggerEndpoint("/swagger/v1/swagger.json", "Arcus Event Grid v1 API");
-                swaggerUiOptions.DisplayOperationId();
-                swaggerUiOptions.EnableDeepLinking();
-                swaggerUiOptions.DocumentTitle = "Arcus Event Grid API";
-                swaggerUiOptions.DocExpansion(DocExpansion.List);
-                swaggerUiOptions.DisplayRequestDuration();
-                swaggerUiOptions.EnableFilter();
-            });
-        }
-
-        private static string GetXmlDocumentationPath(IServiceCollection services)
-        {
-            var hostingEnvironment = services.FirstOrDefault(service => service.ServiceType == typeof(IHostingEnvironment));
-            if (hostingEnvironment == null)
-            {
-                return string.Empty;
-            }
-
-            var contentRootPath = ((IHostingEnvironment)hostingEnvironment.ImplementationInstance).ContentRootPath;
-            var xmlDocumentationPath = $"{contentRootPath}/Open-Api-Docs.xml";
-
-            return File.Exists(xmlDocumentationPath) ? xmlDocumentationPath : string.Empty;
+            app.UseOpenApiDocsWithExplorer();
         }
     }
 }
